+++ Stripe/PublicHeaders/Stripe.h
@@ -38,7 +38,6 @@
 #import "STPImageLibrary.h"
 #import "STPIntentAction.h"
 #import "STPIntentActionRedirectToURL.h"
-#import "STPLegalEntityParams.h"
 #import "STPPaymentActivityIndicatorView.h"
 #import "STPPaymentCardTextField.h"
 #import "STPPaymentConfiguration.h"